Remove ambiguity regarding "drafted" vs "acquired"
Tweak the wording of "Player Salaries" and "Restricted Free Agency" to
make it explicitly clear that a player's salary or RFA eligibility is
determined based on whether or not they were drafted and in which draft
they were selected.   The method by which a player was acquired by the
current owner doesn't impact the player's salary or their RFA elibility.